# Improve constructor performance