Performance tested on Julia 1.0. 

Before:
```julia
julia> for i in 1:8
          @btime FixedDecimal{Int,$i}.(1:100)
       end
  8.986 μs (3 allocations: 960 bytes)
  9.034 μs (3 allocations: 960 bytes)
  10.391 μs (3 allocations: 960 bytes)
  9.667 μs (3 allocations: 960 bytes)
  10.601 μs (3 allocations: 960 bytes)
  10.630 μs (3 allocations: 960 bytes)
  10.768 μs (3 allocations: 960 bytes)
  9.423 μs (3 allocations: 960 bytes)
```
After (without pre-computing `max_exp10`):
```julia
julia> for i in 1:8
          @btime FixedDecimal{Int,$i}.(1:100)
       end
  531.746 ns (3 allocations: 960 bytes)
  569.186 ns (3 allocations: 960 bytes)
  1.888 μs (3 allocations: 960 bytes)
  706.234 ns (3 allocations: 960 bytes)
  1.934 μs (3 allocations: 960 bytes)
  2.056 μs (3 allocations: 960 bytes)
  2.153 μs (3 allocations: 960 bytes)
  700.782 ns (3 allocations: 960 bytes)
```
After (with pre-computing `max_exp10`):
```julia
julia> for i in 1:8
          @btime FixedDecimal{Int,$i}.(1:100)
       end
  390.299 ns (3 allocations: 960 bytes)
  426.136 ns (3 allocations: 960 bytes)
  1.875 μs (3 allocations: 960 bytes)
  541.351 ns (3 allocations: 960 bytes)
  1.961 μs (3 allocations: 960 bytes)
  1.966 μs (3 allocations: 960 bytes)
  2.189 μs (3 allocations: 960 bytes)
  533.450 ns (3 allocations: 960 bytes)
```
One caveat is that custom `Integer` types that don't support `typemax` will either need to implement `max_exp10(::Type{MyInt}) = -1` or they will fail when calling the fallback `max_exp10` method.