@omus 2018-10-26
Related: https://github.com/RelationalAI-oss/FixedPointDecimals.jl/pull/1