@omus 2018-10-26
The coverage reduction is just this line being missed:
https://github.com/JuliaMath/FixedPointDecimals.jl/blob/98e65b6eb890d666255fdf57ada0d905bae2ec23/src/FixedPointDecimals.jl#L478
That line is definitely being tested by:
https://github.com/JuliaMath/FixedPointDecimals.jl/blob/98e65b6eb890d666255fdf57ada0d905bae2ec23/test/runtests.jl#L106
So this looks to just be an issue with coverage reporting.