@TotalVerb 2018-10-26
@omus I see. If the compiler is able to do constant folding, does precomputation provide additional benefit?