@NHDaly 2018-10-26
> Im surprised precomputing has an impact considering

I was assuming the first numbers simply included compiling in their timings?