@NHDaly 2018-10-26
So, I opened https://github.com/JuliaMath/FixedPointDecimals.jl/issues/35 to discuss this more. I think traits would probably be a better solution than this `@eval` solution.

BUT in this case, basically I actually think we can avoid all this and just remove the `applicable` call. It doesn't really add anything...