@NHDaly 2018-10-26
So i'm currently proposing this solution: https://github.com/JuliaMath/FixedPointDecimals.jl/pull/34

Of course we need to add tests and stuff like you've done here curtis. Or, if you want, we can keep this PR as the main one, and I can propose changes into it, or w/e you're comfortable with! :) Thanks for taking an interest after reading the Slack thread!! ❤️ !