@omus 2018-10-26
> Im surprised precomputing has an impact

I was also surprised. It seems pretty consistent though.

> I was assuming the first numbers simply included compiling in their timings?

`@btime` runs the functions multiple times and shouldn't be effecting the results that much. I ran all of the above benchmarks with a fresh Julia 1.0 session.