@NHDaly 2018-10-26
> @btime runs the functions multiple times and shouldn't be effecting the results that much. I ran all of the above benchmarks with a fresh Julia 1.0 session.

Oh yeah, duh. Sorry i missed that. It is weird then..