@KristofferC 2018-10-26
The following diff

```diff
diff --git a/src/FixedPointDecimals.jl b/src/FixedPointDecimals.jl
index 385f5df..8debb38 100644
--- a/src/FixedPointDecimals.jl
+++ b/src/FixedPointDecimals.jl
@@ -94,14 +94,17 @@ struct FixedDecimal{T <: Integer, f} <: Real
         if f >= 0 && (n < 0 || f <= n)
             new{T, f}(i % T)
         else
-            throw(ArgumentError(
-                "Requested number of decimal places $f exceeds the max allowed for the " *
-                "storage type $T: [0, $n]"
-            ))
+            __throw_argerror(f, T, n)
         end
     end
 end

+
+@noinline __throw_argerror(f, T, n) = throw(ArgumentError(
+    "Requested number of decimal places $f exceeds the max allowed for the " *
+    "storage type $T: [0, $n]"
+))
+
```

gives a pretty good speedup.