@KristofferC 2018-10-26
I also don't get any speedup from the `@eval` on master.