@NHDaly 2018-10-26
> I also don't get any speedup from the `@eval` on master.

oh weird, really? Are you sure? What do you mean "master" -- master julia?

I definitely do get the speedups from this PR:
```
# This is all after checking out the `cv/constructor-performance` branch.
julia> @code_native reinterpret(FixedPointDecimals.FixedDecimal{UInt8,0}, 1)
    .section    __TEXT,__text,regular,pure_instructions
; Function reinterpret {
; Location: FixedPointDecimals.jl:93
    movl    %esi, %eax
    retl
    nopw    %cs:(%eax,%eax)
;}

julia> _test(10_000_000)
  0.082648 seconds (2 allocations: 76.294 MiB, 86.26% gc time)
  0.005308 seconds
  0.023281 seconds (2 allocations: 76.294 MiB, 47.62% gc time)
  0.004669 seconds
```