@KristofferC 2018-10-26
I meant on master julia I don't see a difference in performance if the

```
for T in Base.BitInteger_types
    @eval max_exp10(::Type{$T}) = $(max_exp10(T))
end
```

bit is taken out.