@NHDaly 2018-10-26
Ah, that's because this PR removes the `applicable` check:
https://github.com/JuliaMath/FixedPointDecimals.jl/pull/30/files#diff-ea4e0198fdaaab3db0fa802fdb33c161L463

And hard-codes the value for `BigInt`:
https://github.com/JuliaMath/FixedPointDecimals.jl/pull/30/files#diff-ea4e0198fdaaab3db0fa802fdb33c161R478

So for a new custom type that doesn't define a `typemax`, it will `MethodError`, whereas before this PR, a custom type would get the `-1` definition.