@NHDaly 2018-10-26
> whereas before this PR, a custom type would get the -1 definition.

So maybe that custom definition for `BigInt` could just be changed to a default definition for all types?