@omus 2018-10-26
Odd I'm not seeing any speed up with the `__throw_argerror` change. I'm on Julia 1.0.0. I'll try updating.