@omus 2018-10-26
> gives a pretty good speedup.

@KristofferC can you share your benchmark and results?