@KristofferC 2018-10-26
Heh, so apparently there is only a performance improvement if the 

```
for T in Base.BitInteger_types
    @eval max_exp10(::Type{$T}) = $(max_exp10(T))
end
```

part is removed. Otherwise I guess that the throw part is optimized away.  In other words, the following has the same performance as this PR

```diff
--- a/src/FixedPointDecimals.jl
+++ b/src/FixedPointDecimals.jl
@@ -94,14 +94,15 @@ struct FixedDecimal{T <: Integer, f} <: Real
         if f >= 0 && (n < 0 || f <= n)
             new{T, f}(i % T)
         else
-            throw(ArgumentError(
-                "Requested number of decimal places $f exceeds the max allowed for the " *
-                "storage type $T: [0, $n]"
-            ))
+            __throw_arg(n, f, T)
         end
     end
 end

+@noinline __throw_arg(n ,f, T) = throw(ArgumentError(
+                "Requested number of decimal places $f exceeds the max allowed for the " *
+                "storage type $T: [0, $n]"
+            ))
 const FD = FixedDecimal

 (::Type{T})(x::Real) where {T <: FD} = convert(T, x)
@@ -477,10 +478,6 @@ end

 max_exp10(::Type{BigInt}) = -1

-# Pre-compute max_exp10 for the standard Integer types
-for T in Base.BitInteger_types
-    @eval max_exp10(::Type{$T}) = $(max_exp10(T))
-end
```